Feat: 게시물 관련 API 및 컴포넌트 리팩토링, 클라이언트와 서버 요청 분리
