+++ src/features/community/api/getPostsServer.ts
@@ -1,21 +1,6 @@
-import { clientFetcher, ClientFetcherOptions } from "@/shared/api/clientFetcher";
 import { serverFetcher, ServerFetcherOptions } from "@/shared/api/serverFetcher";
 
 import { GetPostsParams, GetPostsResponse } from "./type";
-
-function buildClientPostsRequestOptions(params: GetPostsParams): ClientFetcherOptions {
-  const { view = "main", after, limit } = params;
-
-  return {
-    method: "GET",
-    auth: "public",
-    query: {
-      view,
-      after,
-      limit,
-    },
-  };
-}
 
 function buildServerPostsRequestOptions(params: GetPostsParams): ServerFetcherOptions {
   const { view = "main", after, limit } = params;
@@ -29,12 +14,6 @@ function buildServerPostsRequestOptions(params: GetPostsParams): ServerFetcherOp
       limit,
     },
   };
-}
-
-export async function getPostsClient(params: GetPostsParams): Promise<GetPostsResponse> {
-  const options = buildClientPostsRequestOptions(params);
-
-  return clientFetcher<GetPostsResponse>("/community/posts", options);
 }
 
 export async function getPostsServer(params: GetPostsParams): Promise<GetPostsResponse> {
